refactor: hook -> hooks 로 변경
